Factor out recursion guard helper type utils into a new type_utils.ts file.
